+++ sdk/security/keyvault/azkeys/models_serde.go
@@ -399,7 +399,7 @@ func (k KeyAttributes) MarshalJSON() ([]byte, error) {
 	populate(objectMap, "enabled", k.Enabled)
 	populateTimeUnix(objectMap, "exp", k.Expires)
 	populate(objectMap, "exportable", k.Exportable)
-	populate(objectMap, "hsmPlatform", k.HsmPlatform)
+	populate(objectMap, "HSMPlatform", k.HSMPlatform)
 	populateTimeUnix(objectMap, "nbf", k.NotBefore)
 	populate(objectMap, "recoverableDays", k.RecoverableDays)
 	populate(objectMap, "recoveryLevel", k.RecoveryLevel)
@@ -428,8 +428,8 @@ func (k *KeyAttributes) UnmarshalJSON(data []byte) error {
 		case "exportable":
 			err = unpopulate(val, "Exportable", &k.Exportable)
 			delete(rawMsg, key)
-		case "hsmPlatform":
-			err = unpopulate(val, "HsmPlatform", &k.HsmPlatform)
+		case "HSMPlatform":
+			err = unpopulate(val, "HSMPlatform", &k.HSMPlatform)
 			delete(rawMsg, key)
 		case "nbf":
 			err = unpopulateTimeUnix(val, "NotBefore", &k.NotBefore)